+++ src/Sulu/Bundle/Sales/OrderBundle/Tests/Functional/Controller/OrderControllerTest.php
@@ -11,6 +11,8 @@
 namespace Sulu\Bundle\Sales\OrderBundle\Tests\Functional\Controller;
 
 use DateTime;
+use Sulu\Bundle\Sales\CoreBundle\Entity\Item;
+use Sulu\Bundle\Sales\CoreBundle\Entity\ItemInterface;
 use Sulu\Bundle\Sales\OrderBundle\Entity\OrderStatus;
 use Sulu\Bundle\Sales\OrderBundle\Tests\OrderTestBase;
 
@@ -301,19 +303,41 @@ class OrderControllerTest extends OrderTestBase
                     'product' => [
                         'id' => $this->data->product->getId()
                     ]
-                ]
+                ],
+                [
+                    'name' => $this->data->productTranslation->getName(),
+                    'number' => $this->data->addon->getAddon()->getNumber(),
+                    'quantity' => 2,
+                    'quantityUnit' => 'pc',
+                    'price' => 1000,
+                    'discount' => 10,
+                    'tax' => 20,
+                    'description' => $this->data->productTranslation->getLongDescription(),
+                    'useProductsPrice' => false,
+                    'weight' => 12,
+                    'width' => 13,
+                    'height' => 14,
+                    'length' => 15,
+                    'supplierName' => 'supplier',
+                    'isRecurringPrice' => true,
+                    'type' => Item::TYPE_ADDON,
+                    'addon' => [
+                        'id' => $this->data->addon->getId()
+                    ]
+                ],
             ]
         ];
         $client = $this->createAuthenticatedClient();
 
         $client->request('POST', '/api/orders', $data);
         $response = json_decode($client->getResponse()->getContent());
-        $this->assertEquals(200, $client->getResponse()->getStatusCode(), $response);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode(), $client->getResponse());
 
         $client->request('GET', '/api/orders/' . $response->id);
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
 
         $this->assertTrue($response->items[0]->isRecurringPrice);
+        $this->assertEquals(Item::TYPE_ADDON, $response->items[1]->type);
     }
 
 
@@ -355,6 +379,7 @@ class OrderControllerTest extends OrderTestBase
                     'price' => 1000,
                     'discount' => 10,
                     'tax' => 20,
+                    'type' => Item::TYPE_CUSTOM,
                     'description' => $this->data->productTranslation->getLongDescription(),
                 ]
             ]
@@ -363,7 +388,7 @@ class OrderControllerTest extends OrderTestBase
 
         $client->request('POST', '/api/orders', $data);
         $response = json_decode($client->getResponse()->getContent());
-        $this->assertEquals(200, $client->getResponse()->getStatusCode(), $response);
+        $this->assertEquals(200, $client->getResponse()->getStatusCode(), $client->getResponse());
 
         $client->request('GET', '/api/orders/' . $response->id);
         $this->assertEquals(200, $client->getResponse()->getStatusCode());